requirements.txt
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==6.0.1
attrs==25.3.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
